Refactor and implement JWT authentication and request body validation filters
