+++ api-gateway3/src/main/java/org/eclipse/ecsp/gateway/rest/ApiGatewayController.java
@@ -23,7 +23,6 @@ import org.eclipse.ecsp.gateway.model.Health;
 import org.eclipse.ecsp.gateway.service.IgniteRouteLocator;
 import org.eclipse.ecsp.utils.logger.IgniteLogger;
 import org.eclipse.ecsp.utils.logger.IgniteLoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
