+++ api-gateway3/src/main/java/org/eclipse/ecsp/gateway/service/ApiRoutesRefreshScheduler.java
@@ -20,7 +20,6 @@ package org.eclipse.ecsp.gateway.service;
 
 import org.eclipse.ecsp.utils.logger.IgniteLogger;
 import org.eclipse.ecsp.utils.logger.IgniteLoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.EnableAsync;
