+++ api-gateway3/src/test/java/org/eclipse/ecsp/gateway/plugins/RequestBodyFilterTest.java
@@ -21,6 +21,7 @@ package org.eclipse.ecsp.gateway.plugins;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.eclipse.ecsp.gateway.exceptions.ApiGatewayException;
+import org.eclipse.ecsp.gateway.plugins.filters.RequestBodyFilter;
 import org.eclipse.ecsp.gateway.utils.GatewayConstants;
 import org.eclipse.ecsp.gateway.utils.ObjectMapperUtil;
 import org.junit.jupiter.api.Assertions;
