+++ api-registry-common/src/test/java/org/eclipse/ecsp/restclient/RestTemplateErrorHandlerTest.java
@@ -97,6 +97,7 @@ class RestTemplateErrorHandlerTest {
 
             @Override
             public void close() {
+                // No implementation needed
             }
 
             @Override
